Moved CLI modal icon next to run job button
